Adiconei mais informção no codigo

+++ aula03.py
@@ -1,4 +1,4 @@
 nome = input("Informe seu nome: ")
-idade = float(input("infome sua idade: "))
+idade = int(input("infome sua idade: "))
 salario = float(input("Informe seu salario: "))
 print(f"Seu nome é {nome}, você tem {idade}anos e seu salario é {salario}")
